Added RTN3. Made the Fake transport behave closer to the real one
